# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai:generateImage:
        post:
            tags:
                - AIService
            description: GenerateImage generates an image with the instance's managed image model.
            operationId: AIService_GenerateImage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenerateImageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GenerateImageResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GenerateImageRequest:
            required:
                - prompt
            type: object
            properties:
                prompt:
                    type: string
                    description: Required. Natural-language description of the image to generate.
                aspectRatio:
                    type: string
                    description: 'Optional. Output aspect ratio. Supported values: 1:1, 4:3, 3:4, 16:9, 9:16.'
        GenerateImageResponse:
            type: object
            properties:
                content:
                    type: string
                    description: Generated image bytes.
                    format: bytes
                contentType:
                    type: string
                    description: MIME type of the generated image.
                filename:
                    type: string
                    description: Suggested filename for attaching the image to a memo.
                model:
                    type: string
                    description: Exact Anyhost Catalog model ID that generated the image.
                generationId:
                    type: string
                    description: Anyhost Generation ID for usage and settlement lookup.
                settlementStatus:
                    type: string
                    description: Settlement state returned by Anyhost.
                cost:
                    type: string
                    description: Final USD cost. Populated only when settlement_status is "settled".
                credits:
                    type: string
                    description: Final Model Credit debit. Populated only when settlement_status is "settled".
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
        TranscribeResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The transcribed text.
        TranscriptionAudio:
            type: object
            properties:
                content:
                    writeOnly: true
                    type: string
                    description: Inline audio bytes.
                    format: bytes
                uri:
                    type: string
                    description: URI for audio content. Reserved for future use.
                filename:
                    type: string
                    description: Optional. The uploaded filename.
                contentType:
                    type: string
                    description: Optional. The MIME type of the input audio.
...